refactor: extract App.tsx logic into focused custom hooks
Move complex state management out of App.tsx into dedicated hooks:
- useEnvironment: environment check, envStatus, isReady
- useOpenClawUpdate: OpenClaw update check/install flow
- useManagerUpdate: Manager (Tauri) self-update flow
- useSecurityCheck: version security warning

Also move shared IPC types (EnvironmentStatus, UpdateInfo, UpdateResult,
SecureVersionInfo) from App.tsx into src/lib/tauri.ts per project conventions,
and update Dashboard/index.tsx to import EnvironmentStatus from tauri.ts.
